Refactor: move markdown code into its own file
This file is clearly way too long, so this is a first step.

Moved the markdown related code into its own file, and the test
file for it into a 'tests' directory inside the nofos app.
